Remove unused class in dolphinscheduler-service
