+++ dolphinscheduler-service/src/main/java/org/apache/dolphinscheduler/service/command/CommandServiceImpl.java
@@ -18,28 +18,18 @@
 package org.apache.dolphinscheduler.service.command;
 
 import static org.apache.dolphinscheduler.common.constants.CommandKeyConstants.CMD_PARAM_RECOVER_WORKFLOW_ID_STRING;
-import static org.apache.dolphinscheduler.common.constants.CommandKeyConstants.CMD_PARAM_SUB_WORKFLOW_DEFINITION_CODE;
 
 import org.apache.dolphinscheduler.common.constants.Constants;
 import org.apache.dolphinscheduler.common.enums.CommandType;
-import org.apache.dolphinscheduler.common.enums.TaskDependType;
 import org.apache.dolphinscheduler.common.utils.JSONUtils;
 import org.apache.dolphinscheduler.dao.entity.Command;
 import org.apache.dolphinscheduler.dao.entity.ErrorCommand;
 import org.apache.dolphinscheduler.dao.entity.Schedule;
-import org.apache.dolphinscheduler.dao.entity.TaskInstance;
-import org.apache.dolphinscheduler.dao.entity.WorkflowDefinition;
-import org.apache.dolphinscheduler.dao.entity.WorkflowInstance;
-import org.apache.dolphinscheduler.dao.entity.WorkflowInstanceRelation;
 import org.apache.dolphinscheduler.dao.mapper.CommandMapper;
 import org.apache.dolphinscheduler.dao.mapper.ErrorCommandMapper;
 import org.apache.dolphinscheduler.dao.mapper.ScheduleMapper;
 import org.apache.dolphinscheduler.dao.mapper.WorkflowDefinitionMapper;
-import org.apache.dolphinscheduler.plugin.task.api.enums.Direct;
-import org.apache.dolphinscheduler.plugin.task.api.model.Property;
-import org.apache.dolphinscheduler.service.utils.ParamUtils;
 
-import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 
 import java.util.EnumMap;
@@ -134,69 +124,4 @@ public class CommandServiceImpl implements CommandService {
         return isNeedCreate;
     }
 
-    @Override
-    public Command createSubProcessCommand(WorkflowInstance parentWorkflowInstance, WorkflowInstance childInstance,
-                                           WorkflowInstanceRelation instanceMap, TaskInstance task) {
-        CommandType commandType = getSubCommandType(parentWorkflowInstance, childInstance);
-        Map<String, Object> subProcessParam = JSONUtils.toMap(task.getTaskParams(), String.class, Object.class);
-        long childDefineCode = 0L;
-        if (subProcessParam.containsKey(CMD_PARAM_SUB_WORKFLOW_DEFINITION_CODE)) {
-            try {
-                childDefineCode =
-                        Long.parseLong(
-                                String.valueOf(subProcessParam.get(CMD_PARAM_SUB_WORKFLOW_DEFINITION_CODE)));
-            } catch (NumberFormatException nfe) {
-                log.error("processDefinitionCode is not a number", nfe);
-                return null;
-            }
-        }
-        WorkflowDefinition subWorkflowDefinition = processDefineMapper.queryByCode(childDefineCode);
-
-        Object localParams = subProcessParam.get(Constants.LOCAL_PARAMS);
-        List<Property> allParam = JSONUtils.toList(JSONUtils.toJsonString(localParams), Property.class);
-        Map<String, String> globalMap = ParamUtils.getGlobalParamMap(task.getVarPool());
-        Map<String, String> fatherParams = new HashMap<>();
-        if (CollectionUtils.isNotEmpty(allParam)) {
-            for (Property info : allParam) {
-                if (Direct.OUT == info.getDirect()) {
-                    continue;
-                }
-                fatherParams.put(info.getProp(), globalMap.get(info.getProp()));
-            }
-        }
-        String processParam = ParamUtils.getSubWorkFlowParam(instanceMap, parentWorkflowInstance, fatherParams);
-        int subProcessInstanceId =
-                childInstance == null ? 0 : (childInstance.getId() == null ? 0 : childInstance.getId());
-        return new Command(
-                commandType,
-                TaskDependType.TASK_POST,
-                parentWorkflowInstance.getFailureStrategy(),
-                parentWorkflowInstance.getExecutorId(),
-                subWorkflowDefinition.getCode(),
-                processParam,
-                parentWorkflowInstance.getWarningType(),
-                parentWorkflowInstance.getWarningGroupId(),
-                parentWorkflowInstance.getScheduleTime(),
-                task.getWorkerGroup(),
-                task.getEnvironmentCode(),
-                parentWorkflowInstance.getWorkflowInstancePriority(),
-                parentWorkflowInstance.getDryRun(),
-                subProcessInstanceId,
-                subWorkflowDefinition.getVersion(),
-                parentWorkflowInstance.getTestFlag());
-    }
-
-    /**
-     * get sub work flow command type
-     * child instance exist: child command = fatherCommand
-     * child instance not exists: child command = fatherCommand[0]
-     */
-    private CommandType getSubCommandType(WorkflowInstance parentWorkflowInstance, WorkflowInstance childInstance) {
-        CommandType commandType = parentWorkflowInstance.getCommandType();
-        if (childInstance == null) {
-            String fatherHistoryCommand = parentWorkflowInstance.getHistoryCmd();
-            commandType = CommandType.valueOf(fatherHistoryCommand.split(Constants.COMMA)[0]);
-        }
-        return commandType;
-    }
 }
