--- dolphinscheduler-service/src/main/java/org/apache/dolphinscheduler/service/utils/ProcessData.java
@@ -1,47 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.dolphinscheduler.service.utils;
-
-import org.apache.dolphinscheduler.plugin.task.api.model.Property;
-import org.apache.dolphinscheduler.service.model.TaskNode;
-
-import java.util.List;
-
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-public class ProcessData {
-
-    @EqualsAndHashCode.Include
-    private List<TaskNode> tasks;
-
-    @EqualsAndHashCode.Include
-    private List<Property> globalParams;
-
-    private int timeout;
-
-    private int tenantId;
-
-    public ProcessData(List<TaskNode> tasks, List<Property> globalParams) {
-        this.tasks = tasks;
-        this.globalParams = globalParams;
-    }
-}
